gpt2_linear_regression/requirements.txt
# Core dependencies
jax>=0.4.0
jaxlib>=0.4.0
flax>=0.7.0
optax>=0.1.0
transformers>=4.30.0

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Utilities
numpy>=1.21.0
tqdm>=4.64.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0
mypy>=0.991